Scaffold plays and shift JS bundles as necessary
