+++ src/_layouts/home.html
@@ -7,6 +7,4 @@ layout: default
 
   {{ content }}
 
-  <script type="text/javascript" src="/assets/bundle.js" charset="utf-8"></script>
-
 </div>
