+++ src/_layouts/play.html
@@ -1,0 +1,12 @@
+---
+layout: page
+---
+
+<div class="play">
+
+  {{ content }}
+  <div id="play"></div>
+
+</div>
+
+<script type="text/javascript" src="/assets/play.bundle.js" charset="utf-8"></script>
