+++ src/_plays/hashitomi.md
@@ -1,0 +1,4 @@
+---
+layout: play
+title: Hashitomi
+---
