+++ src/_plays/kokaji.md
@@ -1,0 +1,4 @@
+---
+layout: play
+title: Kokaji
+---
